implementación de la realización de multiples pedidos y mejora de estilos

+++ clientereact/src/components/buttons/Dinner/DetailsButton.js
@@ -5,6 +5,8 @@ import axios from "axios";
 
 export function DetailsButton(props) {
     const navigate = useNavigate();
+
+    const [articles, setArticles] = useState([]);
     const owner = useLocation().pathname.split('/')[1];
     const codeNumber = useLocation().pathname.split('/')[3];
 
@@ -16,18 +18,27 @@ export function DetailsButton(props) {
 
     const order = async () => {
         try {
-            await Promise.all(props.articles.map(async (article, index) => {
-                await axios.post('http://172.17.0.2:8888/order_log', {
-                    owner_name: props.idOwner.name,
-                    article: article.name,
-                    price: article.price,
-                    number_table: props.table,
-                    codetable: codeNumber,
-                    quantity: props.numberArticles[index]
-                });
-            }));
-            await axios.delete(`http://172.17.0.2:8888/delete_basket/${props.idOwner.id_restaurant}/${props.table}`);
-            navigate(`/${owner}/pedido/${codeNumber}`);
+            const resumeOrderResponse = await axios.get(`http://172.17.0.2:8888/resumeOrder/${props.idOwner.id_restaurant}/${props.table}`);
+            setArticles(resumeOrderResponse.data);
+
+            if (resumeOrderResponse.data.length > 0){
+                await Promise.all(resumeOrderResponse.data.map(async (article, index) => {
+                    await axios.post('http://172.17.0.2:8888/order_log', {
+                        owner_name: props.idOwner.name,
+                        article: article.name,
+                        price: article.price,
+                        number_table: props.table,
+                        codetable: codeNumber,
+                        quantity: props.numberArticles[index]
+                    });
+                }));
+                await axios.delete(`http://172.17.0.2:8888/delete_basket/${props.idOwner.id_restaurant}/${props.table}`);
+                navigate(`/${owner}/pedido/${codeNumber}`);
+                localStorage.setItem('pedido','2');
+            }else {
+                navigate(`/${owner}/pedido/${codeNumber}`);
+                localStorage.setItem('pedido','3');
+            }
         } catch (e) {
             console.error("Error al realizar el pedido: " + e);
         }
